Change ticks parameter type from Integer to Long in Source and SourceRepository for improved precision

+++ src/main/java/es/vodafone/sid/poller/model/Source.java
@@ -29,7 +29,7 @@ public record Source(
     BigInteger cache,
     Short scale,
     Boolean active,
-    Integer ticks
+    Long ticks
 ) {
   public boolean isMulti() {
     return SourceTypeRegistry.isMulti(type);
